requirements.txt
albumentations==2.0.8
ants==0.0.7
antspyx==0.6.1
einops==0.8.1
matplotlib==3.10.7
ml_collections==1.1.0
natsort==8.4.0
nibabel==5.3.2
numpy==2.3.5
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==12.0.0
pydicom==3.0.1
pystrum==0.4
scikit_learn==1.7.2
scipy==1.16.3
skimage==0.0
tensorflow==2.20.0
tf==1.1.0
timm==1.0.22
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
wandb==0.23.0
